pyproject.toml
[project]
name = "sttcli"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "rich>=13.7",
    "ffmpeg-python>=0.2.0",
    "openai-whisper>=20231117",
    "openai>=1.30",
    "google-genai>=0.8",
    "elevenlabs>=1.13",
]

[project.scripts]
sttcli = "sttcli.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sttcli"]
